Fixed issue with parseTrace() service
