+++ internal/iso/data/data_set_manager.go
@@ -1,4 +1,4 @@
-package server
+package data
 
 import (
 	"bytes"
