--- internal/services/v0/handlers/isoserver/active_servers.go
@@ -1,16 +1,0 @@
-package isoserver
-
-import (
-	"isosim/internal/iso/server"
-	"net/http"
-)
-
-func addGetActiveServersHandler() {
-
-	http.HandleFunc("/iso/v0/server/active", func(rw http.ResponseWriter, req *http.Request) {
-
-		data := server.ActiveServers()
-		_, _ = rw.Write([]byte(data))
-
-	})
-}
