+++ internal/services/v0/handlers/isoserver/isoserv.go
@@ -2,9 +2,7 @@ package isoserver
 
 import (
 	log "github.com/sirupsen/logrus"
-	"isosim/internal/iso"
 	"net/http"
-	"path/filepath"
 )
 
 func AddAll() {
@@ -15,22 +13,6 @@ func AddAll() {
 	startServerHandler()
 	addGetActiveServersHandler()
 	stopServerHandler()
-
-}
-
-func addIsoServerHandlers() {
-
-	log.Print("Adding ISO server handler .. ")
-	http.HandleFunc("/iso/v0/server", func(rw http.ResponseWriter, req *http.Request) {
-
-		pattern := "/iso/v0/server"
-		log.Debugf("Pattern: %s . Requested URI = %s", pattern, req.RequestURI)
-
-		file := filepath.Join(iso.HTMLDir, "iso_server.html")
-		log.Debugln("Serving file = " + file)
-		http.ServeFile(rw, req, file)
-
-	})
 
 }
