--- internal/services/v0/handlers/isosim_urls.go
@@ -1,13 +1,0 @@
-package handlers
-
-const (
-	homeUrl               = "/iso/home"
-	AllSpecsUrl           = "/iso/v0/specs"
-	SpecMessagesUrl       = "/iso/v0/msgs/"     //{specId}
-	MessageTemplateUrl    = "/iso/v0/template/" //{specId}/{msgId}
-	ParseTraceUrl         = "/iso/v0/parse/"    //{specId}/{msgId}/
-	ParseTraceExternalUrl = "/iso/v0/parse/external"
-	SendMsgUrl            = "/iso/v0/send"
-	SaveMsgUrl            = "/iso/v0/save"
-	LoadMsgUrl            = "/iso/v0/loadmsg"
-)
