+++ internal/services/websim/http_urls.go
@@ -5,7 +5,7 @@ const (
 	URLMessages4Spec = "/iso/v1/msgs/" //{specId}
 
 	URLGetMessageTemplate = "/iso/v1/template/" //{specId}/{msgId}
-	URLParseTrace         = "/iso/v5/parse"     //{specId}/{msgId}/
+	URLParseTrace         = "/iso/v1/parse/"    //{specId}/{msgId}/
 	URLParseTraceExternal = "/iso/v1/parse/external"
 	URLSendMessageToHost  = "/iso/v1/send"
 	URLSaveMsg            = "/iso/v1/save"
